+++ custom_components/electrolux_ac/sensor.py
@@ -132,7 +132,7 @@ class TemperatureSensor(SensorBase):
 
     _LOGGER.debug("Creating temperature sensor with presentation: %s",
                   self._appliance._states.get('temperatureRepresentation'))
-    if self._appliance._states.get('temperatureRepresentation') == 'fahrenheit':
+    if (self._appliance._states.get('temperatureRepresentation') or '').lower() == 'fahrenheit':
       self._attr_native_unit_of_measurement = UnitOfTemperature.FAHRENHEIT
     else:
       self._attr_native_unit_of_measurement = UnitOfTemperature.CELSIUS